json.c: move StringCaptureData() from regex.c
