+++ libutils/json.c
@@ -34,6 +34,7 @@
 #include <file_lib.h>
 #include <printsize.h>
 #include <regex.h>
+#include <buffer.h>
 
 static const int SPACES_PER_INDENT = 2;
 const int DEFAULT_CONTAINER_CAPACITY = 64;
@@ -2332,3 +2333,46 @@ JsonParseError JsonParseFile(const char *path, size_t size_max, JsonElement **js
 {
     return JsonParseAnyFile(path, size_max, json_out, false);
 }
+
+/*******************************************************************/
+
+// returns NULL on any failure
+// takes either a pre-compiled pattern OR a regex (one of the two shouldn't be NULL)
+JsonElement* StringCaptureData(pcre *pattern, const char* regex, const char* data)
+{
+    assert(regex || pattern);
+    assert(data);
+
+    Seq *s;
+
+    if (NULL != pattern)
+    {
+        s = StringMatchCapturesWithPrecompiledRegex(pattern, data, true);
+    }
+    else
+    {
+        s = StringMatchCaptures(regex, data, true);
+    }
+
+    if (!s || SeqLength(s) == 0)
+    {
+        SeqDestroy(s);
+        return NULL;
+    }
+
+    JsonElement *json = JsonObjectCreate(SeqLength(s)/2);
+
+    for (int i = 1; i < SeqLength(s); i+=2)
+    {
+        Buffer *key = SeqAt(s, i-1);
+        Buffer *value = SeqAt(s, i);
+
+        JsonObjectAppendString(json, BufferData(key), BufferData(value));
+    }
+
+    SeqDestroy(s);
+
+    JsonObjectRemoveKey(json, "0");
+    return json;
+}
+
