+++ libutils/regex.h
@@ -32,7 +32,6 @@
 #include <pcre_include.h>
 
 #include <sequence.h>                                           /* Seq */
-#include <json.h>
 
 #define CFENGINE_REGEX_WHITESPACE_IN_CONTEXTS ".*[_A-Za-z0-9][ \\t]+[_A-Za-z0-9].*"
 
@@ -45,7 +44,6 @@ bool StringMatchFull(const char *regex, const char *str);
 bool StringMatchFullWithPrecompiledRegex(pcre *regex, const char *str);
 Seq *StringMatchCaptures(const char *regex, const char *str, const bool return_names);
 Seq *StringMatchCapturesWithPrecompiledRegex(const pcre *pattern, const char *str, const bool return_names);
-JsonElement* StringCaptureData(pcre *pattern, const char* regex, const char* data);
 bool CompareStringOrRegex(const char *value, const char *compareTo, bool regex);
 
 #endif  /* CFENGINE_REGEX_H */
